requirements.txt
certifi==2022.12.7
click==8.1.3
colorama==0.4.4
cycler==0.11.0
dnspython==2.2.1
Flask==2.2.3
gunicorn==20.1.0
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
kiwisolver==1.3.1
MarkupSafe==2.1.1
matplotlib==3.4.2
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.16.0
Werkzeug==2.1.1
wincertstore==0.2
zipp==3.8.0
Pyrebase4==4.6.0
PyRect==0.2.0
pyrsistent==0.19.3
PyScreeze==0.1.28
Flask-RESTful==0.3.9